Factor out verifyApiToken() tests
